Registration system update, main form update

+++ MiniSteam.Data/Game.cs
@@ -8,7 +8,7 @@ namespace MiniSteam.Data
 {
     public sealed class Game
     {
-        public float Rating { get; set;}
+        public int Rating { get; set;}
         public string GameName { get; set;}
         public string Description { get; set;}
         public string Author { get; set; }
